Add provisioner file ops, and add permissions checking for them.

+++ go/provisioner-mgmt/isos.go
@@ -13,12 +13,17 @@ import (
 )
 
 func listIsos(c *gin.Context, fileRoot string) {
+	user := getUser(c)
+	res := []string{}
+	if !testCap(c, int(user.TenantID), "PROVISIONER_FILE_READ") {
+		c.JSON(http.StatusOK, res)
+		return
+	}
 	ents, err := ioutil.ReadDir(path.Join(fileRoot, "isos"))
 	if err != nil {
 		c.JSON(http.StatusNotFound, NewError(fmt.Sprintf("list: error listing isos: %v", err)))
 		return
 	}
-	res := []string{}
 	for _, ent := range ents {
 		if !ent.Mode().IsRegular() {
 			continue
@@ -29,6 +34,11 @@ func listIsos(c *gin.Context, fileRoot string) {
 }
 
 func getIso(c *gin.Context, fileRoot, name string) {
+	user := getUser(c)
+	if !testCap(c, int(user.TenantID), "PROVISIONER_FILE_READ") {
+		c.Data(http.StatusNotFound, "application/octet-stream", nil)
+		return
+	}
 	isoName := path.Join(fileRoot, `isos`, path.Base(name))
 	c.File(isoName)
 }
@@ -50,8 +60,14 @@ func reloadBootenvsForIso(name string) {
 }
 
 func uploadIso(c *gin.Context, fileRoot, name string) {
+	user := getUser(c)
+	if !testCap(c, int(user.TenantID), "PROVISIONER_FILE_CREATE") {
+		c.JSON(http.StatusForbidden, NewError("upload: access denied"))
+		return
+	}
 	if c.Request.Header.Get(`Content-Type`) != `application/octet-stream` {
 		c.JSON(http.StatusUnsupportedMediaType, NewError(fmt.Sprintf("upload: iso %s must have content-type application/octet-stream", name)))
+		return
 	}
 	isoTmpName := path.Join(fileRoot, `isos`, fmt.Sprintf(`.%s.part`, path.Base(name)))
 	isoName := path.Join(fileRoot, `isos`, path.Base(name))
@@ -68,10 +84,12 @@ func uploadIso(c *gin.Context, fileRoot, name string) {
 	if err != nil {
 		os.Remove(isoTmpName)
 		c.JSON(http.StatusInsufficientStorage, NewError(fmt.Sprintf("upload: Failed to upload %s: %v", name, err)))
+		return
 	}
 	if c.Request.ContentLength != 0 && copied != c.Request.ContentLength {
 		os.Remove(isoTmpName)
 		c.JSON(http.StatusBadRequest, NewError(fmt.Sprintf("upload: Failed to upload entire file %s: %d bytes expected, %d bytes recieved", name, c.Request.ContentLength, copied)))
+		return
 	}
 	os.Remove(isoName)
 	os.Rename(isoTmpName, isoName)
@@ -84,9 +102,15 @@ func uploadIso(c *gin.Context, fileRoot, name string) {
 }
 
 func deleteIso(c *gin.Context, fileRoot, name string) {
+	user := getUser(c)
+	if !testCap(c, int(user.TenantID), "PROVISIONER_FILE_DESTROY") {
+		c.JSON(http.StatusForbidden, NewError("delete: access denied"))
+		return
+	}
 	isoName := path.Join(fileRoot, `isos`, path.Base(name))
 	if err := os.Remove(isoName); err != nil {
 		c.JSON(http.StatusNotFound, NewError(fmt.Sprintf("delete: unable to delete %s: %v", name, err)))
+		return
 	}
 	c.JSON(http.StatusAccepted, nil)
 }
